samples/middle-tier/python-fastapi/pyproject.toml
[tool.poetry]
name = "rt-middle-tier"
version = "0.1.0"
description = "FastAPI WebSocket server sample for a realtime middle tier between a front end and a Azure OpenAI or OpenAI backend."
authors = ["Gerardo Lecaros <10088504+glecaros@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.6"
uvicorn = "^0.32.1"
python-dotenv = "^1.0.1"
loguru = "^0.7.3"
azure-identity = "^1.19.0"
azure-core = "^1.32.0"
websockets = "^14.1"
rtclient = { url = "https://github.com/Azure-Samples/aoai-realtime-audio-sdk/releases/download/py%2Fv0.5.3/rtclient-0.5.3.tar.gz" }
langchain-community = "^0.3.25"
soundfile = "^0.12.1"
numpy = "^2.1.0"
scipy = "^1.13.1"
pypdf = "^4.2.0"
docx2txt = "^0.8"
pandas = "^2.2.2"
beautifulsoup4 = "^4.12.3"
markdown = "^3.6"
weaviate-client = "^4.5.5"
docling-core = {extras = ["chunking-openai"], version = "^2.40.0"}
transformers = "^4.53.1"
docling = "^2.40.0"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
ruff = "^0.8.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
